fix(agents): break madge import cycle in harness selection path
