+++ src/auto-reply/reply/commands-status.ts
@@ -1,5 +1,6 @@
 import { logVerbose } from "../../globals.js";
-import { buildStatusText, type BuildStatusTextParams } from "../../status/status-text.js";
+import { buildStatusText } from "../../status/status-text.js";
+import type { BuildStatusTextParams } from "../../status/status-text.types.js";
 import type { ReplyPayload } from "../types.js";
 import type { CommandContext } from "./commands-types.js";
 export { buildStatusText } from "../../status/status-text.js";
